Avoid size_t to int casting in EXPECT/ASSERT/CHECK macros
Using unsigned literals avoids the need for such casts, and reduces the
number of potentially unsafe casts to be investigated down the road.

NO_IFTTT=functionality unchanged.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7218814
Commit-Queue: Nico Weber <thakis@chromium.org>
Owners-Override: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1553821}

+++ ash/system/power/power_notification_controller_unittest.cc
@@ -717,7 +717,7 @@ TEST_P(PowerNotificationControllerWithBatterySaverTest,
         message_center()->FindVisibleNotificationById("battery");
     const std::vector<message_center::ButtonInfo> buttons =
         notification->buttons();
-    EXPECT_EQ(static_cast<int>(buttons.size()), 1);
+    EXPECT_EQ(buttons.size(), 1u);
   }
 
   // Plug in charger.
@@ -760,7 +760,7 @@ TEST_P(PowerNotificationControllerWithBatterySaverTest,
         message_center()->FindVisibleNotificationById("battery");
     const std::vector<message_center::ButtonInfo> buttons =
         notification->buttons();
-    EXPECT_EQ(static_cast<int>(buttons.size()), 1);
+    EXPECT_EQ(buttons.size(), 1u);
 
     // Simulate Clicking Opt-Out/In
     notification->delegate()->Click(0, std::nullopt);
